Refactor: move registration attribute handling to its own module
