+++ services/121-service/src/notifications/message-template/message-template.module.ts
@@ -4,14 +4,14 @@ import { TypeOrmModule } from '@nestjs/typeorm';
 import { MessageTemplateController } from '@121-service/src/notifications/message-template/message-template.controller';
 import { MessageTemplateEntity } from '@121-service/src/notifications/message-template/message-template.entity';
 import { MessageTemplateService } from '@121-service/src/notifications/message-template/message-template.service';
-import { ProgramAttributesModule } from '@121-service/src/program-attributes/program-attributes.module';
+import { ProgramRegistrationAttributesModule } from '@121-service/src/program-registration-attributes/program-registration-attributes.module';
 import { UserModule } from '@121-service/src/user/user.module';
 
 @Module({
   imports: [
     TypeOrmModule.forFeature([MessageTemplateEntity]),
     UserModule,
-    ProgramAttributesModule,
+    ProgramRegistrationAttributesModule,
   ],
   providers: [MessageTemplateService],
   controllers: [MessageTemplateController],
