+++ services/121-service/src/program-registration-attributes/program-registration-attributes.service.ts
@@ -1,0 +1,459 @@
+import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
+import { InjectRepository } from '@nestjs/typeorm';
+import { FilterOperator } from 'nestjs-paginate';
+import { Equal, QueryFailedError, Repository } from 'typeorm';
+
+import {
+  ProgramRegistrationAttributeDto,
+  UpdateProgramRegistrationAttributeDto,
+} from '@121-service/src/programs/dto/program-registration-attribute.dto';
+import { ProgramEntity } from '@121-service/src/programs/entities/program.entity';
+import { ProgramRegistrationAttributeEntity } from '@121-service/src/programs/entities/program-registration-attribute.entity';
+import { ProgramRegistrationAttributeMapper } from '@121-service/src/programs/mappers/program-registration-attribute.mapper';
+import {
+  AllowedFiltersNumber,
+  AllowedFiltersString,
+  PaginateConfigRegistrationView,
+} from '@121-service/src/registration/const/filter-operation.const';
+import { FilterAttributeDto } from '@121-service/src/registration/dto/filter-attribute.dto';
+import {
+  Attribute,
+  GenericRegistrationAttributes,
+  RegistrationAttributeTypes,
+} from '@121-service/src/registration/enum/registration-attribute.enum';
+import { registrationViewAttributeNames } from '@121-service/src/shared/const';
+
+@Injectable()
+export class ProgramRegistrationAttributesService {
+  @InjectRepository(ProgramEntity)
+  private readonly programRepository: Repository<ProgramEntity>;
+  @InjectRepository(ProgramRegistrationAttributeEntity)
+  private readonly programRegistrationAttributeRepository: Repository<ProgramRegistrationAttributeEntity>;
+
+  private async findProgramOrThrow(programId: number): Promise<void> {
+    const program = await this.programRepository.findOne({
+      where: { id: Equal(programId) },
+    });
+    if (!program) {
+      const errors = `No program found with id ${programId}`;
+      throw new HttpException({ errors }, HttpStatus.NOT_FOUND);
+    }
+  }
+
+  public getFilterableAttributes(program: ProgramEntity) {
+    const genericPaAttributeFilters = [
+      'personAffectedSequence',
+      'referenceId',
+      'phoneNumber',
+      'preferredLanguage',
+      'inclusionScore',
+      'paymentAmountMultiplier',
+      'programFspConfigurationName',
+    ];
+    const paAttributesNameArray = program['paTableAttributes'].map(
+      (paAttribute: Attribute) => paAttribute.name,
+    );
+
+    const paymentGroup = {
+      group: 'payments',
+      filters: ['paymentCount'],
+    };
+    if (program.enableMaxPayments) {
+      paymentGroup.filters.push('maxPayments');
+      paymentGroup.filters.push('paymentCountRemaining');
+    }
+
+    const filterableAttributeNames = [
+      paymentGroup,
+      {
+        group: 'messages',
+        filters: ['lastMessageStatus'],
+      },
+      {
+        group: 'paAttributes',
+        filters: [
+          ...new Set([...genericPaAttributeFilters, ...paAttributesNameArray]),
+        ],
+      },
+    ];
+
+    const filterableAttributes: {
+      group: string;
+      filters: FilterAttributeDto[];
+    }[] = [];
+    for (const group of filterableAttributeNames) {
+      const filterableAttributesPerGroup: FilterAttributeDto[] = [];
+      for (const name of group.filters) {
+        if (PaginateConfigRegistrationView.filterableColumns?.[name]) {
+          filterableAttributesPerGroup.push({
+            name,
+            allowedOperators: PaginateConfigRegistrationView.filterableColumns[
+              name
+            ] as FilterOperator[],
+            isInteger:
+              PaginateConfigRegistrationView.filterableColumns[name] ===
+              AllowedFiltersNumber,
+          });
+        } else {
+          // If no allowed operators are defined than the attribute is
+          // registration data which is stored as a string
+          filterableAttributesPerGroup.push({
+            name,
+            allowedOperators: AllowedFiltersString,
+            isInteger: false,
+          });
+        }
+      }
+      filterableAttributes.push({
+        group: group.group,
+        filters: filterableAttributesPerGroup,
+      });
+    }
+
+    return filterableAttributes;
+  }
+
+  public async getAttributes({
+    programId,
+    includeProgramRegistrationAttributes,
+    includeTemplateDefaultAttributes,
+    filterShowInRegistrationsTable,
+  }: {
+    programId: number;
+    includeProgramRegistrationAttributes: boolean;
+    includeTemplateDefaultAttributes: boolean;
+    filterShowInRegistrationsTable?: boolean;
+  }): Promise<Attribute[]> {
+    let programAttributes: Attribute[] = [];
+    if (includeProgramRegistrationAttributes) {
+      programAttributes = await this.getAndMapProgramRegistrationAttributes(
+        programId,
+        filterShowInRegistrationsTable,
+      );
+    }
+
+    let templateDefaultAttributes: Attribute[] = [];
+    if (includeTemplateDefaultAttributes) {
+      templateDefaultAttributes =
+        await this.getMessageTemplateDefaultAttributes(programId);
+    }
+
+    return [...programAttributes, ...templateDefaultAttributes];
+  }
+
+  private async getMessageTemplateDefaultAttributes(
+    programId: number,
+  ): Promise<Attribute[]> {
+    const hasMaxPayments = await this.programRepository.findOne({
+      where: { id: Equal(programId) },
+      select: ['enableMaxPayments'],
+    });
+    const defaultAttributes: Attribute[] = [
+      {
+        name: GenericRegistrationAttributes.paymentAmountMultiplier,
+        type: RegistrationAttributeTypes.numeric,
+        label: null,
+      },
+      {
+        name: GenericRegistrationAttributes.programFspConfigurationLabel,
+        type: RegistrationAttributeTypes.text,
+        label: null,
+      },
+    ];
+    if (hasMaxPayments?.enableMaxPayments) {
+      defaultAttributes.push({
+        name: GenericRegistrationAttributes.maxPayments,
+        type: RegistrationAttributeTypes.numeric,
+        label: null,
+      });
+      defaultAttributes.push({
+        name: GenericRegistrationAttributes.paymentCountRemaining,
+        type: RegistrationAttributeTypes.numeric,
+        label: null,
+      });
+    }
+    return defaultAttributes;
+  }
+
+  public async getPaEditableAttributes(
+    programId: number,
+  ): Promise<Attribute[]> {
+    const programRegistrationAttributes = (
+      await this.programRegistrationAttributeRepository.find({
+        where: { program: { id: Equal(programId) } },
+      })
+    ).map((c) => {
+      return {
+        name: c.name,
+        type: c.type,
+        label: c.label,
+        isRequired: c.isRequired,
+      };
+    });
+    return programRegistrationAttributes;
+  }
+
+  private async getAndMapProgramRegistrationAttributes(
+    programId: number,
+    filterShowInRegistrationsTable?: boolean,
+  ): Promise<Attribute[]> {
+    let queryRegistrationAttr = this.programRegistrationAttributeRepository
+      .createQueryBuilder('programRegistrationAttribute')
+      .orderBy('programRegistrationAttribute.created', 'ASC')
+      .where({ program: { id: programId } });
+
+    if (filterShowInRegistrationsTable) {
+      queryRegistrationAttr = queryRegistrationAttr.andWhere({
+        showInPeopleAffectedTable: true,
+      });
+    }
+    const rawProgramAttributes = await queryRegistrationAttr.getMany();
+    const programAttributes = rawProgramAttributes.map((c) => {
+      return {
+        name: c.name,
+        type: c.type,
+        label: c.label,
+        isRequired: c.isRequired,
+      };
+    });
+
+    return programAttributes;
+  }
+
+  public async applyProgramRegistrationAttributesFallbackIfNecessary({
+    attributesData,
+    namingConventionData,
+  }: {
+    attributesData: ProgramRegistrationAttributeDto[] | undefined;
+    namingConventionData: string[];
+  }): Promise<ProgramRegistrationAttributeDto[]> {
+    const programRegistrationAttributes = attributesData ?? [];
+
+    // make sure phoneNumber is in programRegistrationAttributes
+
+    if (
+      !programRegistrationAttributes.find((attr) => attr.name === 'phoneNumber')
+    ) {
+      programRegistrationAttributes.push({
+        name: 'phoneNumber',
+        type: RegistrationAttributeTypes.text,
+        label: { en: 'Phone number' },
+      });
+    }
+
+    // make sure all fullnameNamingConventions are in programRegistrationAttributes
+
+    const registrationAttributesNames = programRegistrationAttributes.map(
+      (attr) => attr.name,
+    );
+    const missingNamingConventions = namingConventionData.filter(
+      (attr) => !registrationAttributesNames.includes(attr),
+    );
+
+    if (missingNamingConventions.length > 0) {
+      for (const missingNamingConvention of missingNamingConventions) {
+        programRegistrationAttributes.push({
+          name: missingNamingConvention,
+          type: RegistrationAttributeTypes.text,
+          label: {
+            en:
+              missingNamingConvention === 'fullName'
+                ? 'Full name'
+                : missingNamingConvention,
+          },
+        });
+      }
+    }
+
+    return programRegistrationAttributes;
+  }
+
+  public async upsertProgramRegistrationAttributes({
+    programId,
+    programRegistrationAttributes,
+  }: {
+    programId: number;
+    programRegistrationAttributes: ProgramRegistrationAttributeDto[];
+  }): Promise<void> {
+    // Fetch all existing attributes for this program in one query
+    const existingAttributes =
+      await this.programRegistrationAttributeRepository.find({
+        where: { programId: Equal(programId) },
+      });
+
+    const existingAttributesMap = new Map(
+      existingAttributes.map((attr) => [attr.name, attr]),
+    );
+
+    const entitiesToSave: ProgramRegistrationAttributeEntity[] = [];
+
+    for (const attribute of programRegistrationAttributes) {
+      const existingAttribute = existingAttributesMap.get(attribute.name);
+
+      if (existingAttribute) {
+        // Update existing attribute
+        for (const key in attribute) {
+          existingAttribute[key] = attribute[key];
+        }
+        entitiesToSave.push(existingAttribute);
+      } else {
+        // Create new attribute
+        const newAttribute =
+          this.programRegistrationAttributeDtoToEntity(attribute);
+        newAttribute.programId = programId;
+        entitiesToSave.push(newAttribute);
+      }
+    }
+    await this.programRegistrationAttributeRepository.save(entitiesToSave);
+  }
+
+  public async createProgramRegistrationAttribute({
+    programId,
+    createProgramRegistrationAttributeDto,
+  }: {
+    programId: number;
+    createProgramRegistrationAttributeDto: ProgramRegistrationAttributeDto;
+  }): Promise<ProgramRegistrationAttributeDto> {
+    const entity = await this.createProgramRegistrationAttributeEntity({
+      programId,
+      createProgramRegistrationAttributeDto,
+    });
+    return ProgramRegistrationAttributeMapper.entityToDto(entity);
+  }
+
+  private async validateAttributeName(
+    programId: number,
+    name: string,
+  ): Promise<void> {
+    const existingAttributes = await this.getAttributes({
+      programId,
+      includeProgramRegistrationAttributes: true,
+      includeTemplateDefaultAttributes: false,
+    });
+    const existingNames = existingAttributes.map((attr) => {
+      return attr.name;
+    });
+    if (existingNames.includes(name)) {
+      const errors = `Unable to create program registration attribute with name ${name}. The names ${existingNames.join(
+        ', ',
+      )} are already in use`;
+      throw new HttpException({ errors }, HttpStatus.BAD_REQUEST);
+    }
+    if (registrationViewAttributeNames.includes(name)) {
+      const errors = `Unable to create program registration attribute with name ${name}. The names ${registrationViewAttributeNames.join(
+        ', ',
+      )} are forbidden to use`;
+      throw new HttpException({ errors }, HttpStatus.BAD_REQUEST);
+    }
+  }
+
+  public async createProgramRegistrationAttributeEntity({
+    programId,
+    createProgramRegistrationAttributeDto,
+    repository,
+  }: {
+    programId: number;
+    createProgramRegistrationAttributeDto: ProgramRegistrationAttributeDto;
+    repository?: Repository<ProgramRegistrationAttributeEntity>;
+  }): Promise<ProgramRegistrationAttributeEntity> {
+    await this.validateAttributeName(
+      programId,
+      createProgramRegistrationAttributeDto.name,
+    );
+    const programRegistrationAttribute =
+      this.programRegistrationAttributeDtoToEntity(
+        createProgramRegistrationAttributeDto,
+      );
+    programRegistrationAttribute.programId = programId;
+
+    try {
+      if (repository) {
+        return await repository.save(programRegistrationAttribute);
+      } else {
+        return await this.programRegistrationAttributeRepository.save(
+          programRegistrationAttribute,
+        );
+      }
+    } catch (error) {
+      if (error instanceof QueryFailedError) {
+        const errorMessage = error.message; // Get the error message from QueryFailedError
+        throw new HttpException(errorMessage, HttpStatus.BAD_REQUEST);
+      }
+      // Unexpected error
+      throw error;
+    }
+  }
+
+  private programRegistrationAttributeDtoToEntity(
+    dto: ProgramRegistrationAttributeDto,
+  ): ProgramRegistrationAttributeEntity {
+    const programRegistrationAttribute =
+      new ProgramRegistrationAttributeEntity();
+    programRegistrationAttribute.name = dto.name;
+    programRegistrationAttribute.label = dto.label;
+    programRegistrationAttribute.type = dto.type;
+    programRegistrationAttribute.options = dto.options ?? null;
+    programRegistrationAttribute.scoring = dto.scoring ?? {};
+    programRegistrationAttribute.pattern = dto.pattern ?? null;
+    programRegistrationAttribute.editableInPortal =
+      dto.editableInPortal ?? true;
+    programRegistrationAttribute.includeInTransactionExport =
+      dto.includeInTransactionExport ?? false;
+    programRegistrationAttribute.duplicateCheck = dto.duplicateCheck ?? false;
+    programRegistrationAttribute.placeholder = dto.placeholder ?? null;
+    programRegistrationAttribute.isRequired = dto.isRequired ?? false;
+    programRegistrationAttribute.showInPeopleAffectedTable =
+      dto.showInPeopleAffectedTable ?? false;
+    return programRegistrationAttribute;
+  }
+
+  public async updateProgramRegistrationAttribute({
+    programId,
+    programRegistrationAttributeName,
+    updateProgramRegistrationAttribute,
+  }: {
+    programId: number;
+    programRegistrationAttributeName: string;
+    updateProgramRegistrationAttribute: UpdateProgramRegistrationAttributeDto;
+  }): Promise<ProgramRegistrationAttributeEntity> {
+    const programRegistrationAttribute =
+      await this.programRegistrationAttributeRepository.findOne({
+        where: {
+          name: Equal(programRegistrationAttributeName),
+          programId: Equal(programId),
+        },
+      });
+    if (!programRegistrationAttribute) {
+      const errors = `No programRegistrationAttribute found with name ${programRegistrationAttributeName} for program ${programId}`;
+      throw new HttpException({ errors }, HttpStatus.NOT_FOUND);
+    }
+
+    for (const attribute in updateProgramRegistrationAttribute) {
+      programRegistrationAttribute[attribute] =
+        updateProgramRegistrationAttribute[attribute];
+    }
+
+    await this.programRegistrationAttributeRepository.save(
+      programRegistrationAttribute,
+    );
+    return programRegistrationAttribute;
+  }
+
+  public async deleteProgramRegistrationAttribute(
+    programId: number,
+    programRegistrationAttributeId: number,
+  ): Promise<ProgramRegistrationAttributeEntity> {
+    await this.findProgramOrThrow(programId);
+    const programRegistrationAttribute =
+      await this.programRegistrationAttributeRepository.findOne({
+        where: { id: Number(programRegistrationAttributeId) },
+      });
+    if (!programRegistrationAttribute) {
+      const errors = `Program registration attribute with id: '${programRegistrationAttributeId}' not found.'`;
+      throw new HttpException({ errors }, HttpStatus.NOT_FOUND);
+    }
+    return await this.programRegistrationAttributeRepository.remove(
+      programRegistrationAttribute,
+    );
+  }
+}
